User addition! (and other changes)
We finally have users! Currently, users can register with a username, email, and password. However, they can't login with their username. Might add that functionality at a later date, but it is not a priority.

Added a login and register page with appropriate navigation in the navbar that is based on login status. The userForm component changes based on whether user needs to register or login.

Watching authentication is implemented, so we can add route guards if we wish.

Added tab index to certain elements for accessibility.

Added breadcrumbs for more sensible navigation!

Cleaned up some comment code and styled the page to be a bit brighter.

Implemented an alert component, but it is not yet functional.